% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{build_searchstring}
\alias{build_searchstring}
\title{builds the URL search string needed to query datatomb}
\usage{
build_searchstring(
  globalsearch = NA,
  tags = NA,
  name = NA,
  author = NA,
  description = NA,
  after = NA,
  before = NA,
  parentOf = NA,
  hash = NA,
  commit = NA,
  projectname = NA
)
}
\arguments{
\item{tags}{one tag or a vector of tags that must be attached to the dataset}

\item{name}{name of the dataset}

\item{author}{author of the dataset}

\item{description}{searches all descriptions for partial matches}

\item{after}{dataset must have been uploaded after the given date}

\item{before}{dataset must have been uploaded before the given date}

\item{parentOf}{dataset must be a parent of the given hash}

\item{hash}{the dataset's hash must (partially) match this hash}

\item{commit}{searches the generator tags (relevant, e.g., for glacier data sets)}

\item{projectname}{name of the project within the dataset was generated}

\item{global_search}{searches all metadata fields for partial match}
}
\value{
searchstring (URL encoded)
}
\description{
builds the URL search string needed to query datatomb
}
