% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksum.R
\name{check}
\alias{check}
\title{checks if a file matches the given checksum}
\usage{
check(filename, checksum, error = TRUE)
}
\arguments{
\item{filename}{file to check}

\item{checksum}{sha256sum to check against}

\item{error}{if set to false, don't throw but return FALSE.}
}
\value{
TRUE, if the file is correct. FALSE, if the file is incorrect and error=FALSE
}
\description{
checks if a file matches the given checksum
}
