% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{check_and_build_hash}
\alias{check_and_build_hash}
\title{checks if the value is not NA and a hash and returns the corresponding element of the search string}
\usage{
check_and_build_hash(name, value)
}
\arguments{
\item{name}{key-part of the search query}

\item{value}{value-part of the search query}
}
\value{
key-value pair of the form key=value
}
\description{
checks if the value is not NA and a hash and returns the corresponding element of the search string
}
