% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{download a dataset from datatomb}
\usage{
download(hash, file = NA, check_file = TRUE)
}
\arguments{
\item{hash}{the datatomb identifier of a dataset. may also be a unique substring.}

\item{file}{local file target where the dataset will be saved. If NA, the file name of the data set is used.}

\item{check_file}{if set to FALSE, do not check for correctness after download}
}
\value{
path to the stored dataset
}
\description{
download a dataset from datatomb
}
