% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure}
\alias{ensure}
\title{ensures that a file is present and correct}
\usage{
ensure(filename, hash, replace = FALSE)
}
\arguments{
\item{filename}{file that is expected to be here}

\item{hash}{file hash (sha256sum) of the file.}
}
\description{
ensures that a file is present and correct
}
