% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{download metadata from datatomb}
\usage{
metadata(hash)
}
\arguments{
\item{hash}{the datatomb identifier of a dataset. may also be a unique substring.}
}
\value{
list with all available metadata
}
\description{
download metadata from datatomb
}
