% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_hash.R
\name{resolve_hash}
\alias{resolve_hash}
\title{a function that extracts a hash from a list or data frame row}
\usage{
resolve_hash(input)
}
\arguments{
\item{input}{any input (character, list,...) that contains exactly one hash}
}
\value{
character hash
}
\description{
a function that extracts a hash from a list or data frame row
}
