% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_id.R
\name{resolve_id}
\alias{resolve_id}
\title{finds a remote hash to a short id string}
\usage{
resolve_id(partial_hash, no_result_throws = TRUE)
}
\arguments{
\item{no_result_throws}{if set to FALSE, return NA if no such hash exists instead of throwing an error}

\item{query}{search query}
}
\value{
hash or NA
}
\description{
completes a partial hash or otherwise identifying string to a full hash.
If no such hash exists the method throws an error, except no_result_throws=FALSE.
Also throws if more than one hash matches.
}
