% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{search datatomb for datasets}
\usage{
search(
  global_search,
  tags = NA,
  name = NA,
  author = NA,
  description = NA,
  after = NA,
  before = NA,
  parentOf = NA,
  hash = NA,
  commit = NA,
  projectname = NA
)
}
\arguments{
\item{global_search}{searches all metadata fields for partial match}

\item{tags}{one tag or a vector of tags that must be attached to the dataset}

\item{name}{name of the dataset}

\item{author}{author of the dataset}

\item{description}{searches all descriptions for partial matches}

\item{after}{dataset must have been uploaded after the given date}

\item{before}{dataset must have been uploaded before the given date}

\item{parentOf}{dataset must be a parent of the given hash}

\item{hash}{the dataset's hash must (partially) match this hash}

\item{commit}{searches the generator tags (relevant, e.g., for glacier data sets)}

\item{projectname}{name of the project within the dataset was generated}
}
\value{
a tibble / dataframe of search results which includes the name, hash, projectname and the description of the dataset
}
\description{
Searches datatomb for datasets using keywords or free text. Refer to the datatomb documentation for more details and to metadata() if more information on a specific dataset is needed.
}
\examples{
search("dtd.model")
search(tags=c("dtd.model", "macro"), author="mschoen")

}
