% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_token}
\alias{set_token}
\title{set token. validates the token and then sets pkg.env$token}
\usage{
set_token(token)
}
\arguments{
\item{token}{access token as obtained by, e.g., auth.spang-lab.de}
}
\value{
datatomb user name
}
\description{
set token. validates the token and then sets pkg.env$token
}
