% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksum.R
\name{sha256sum}
\alias{sha256sum}
\title{reads a file and computes its sha256sum.}
\usage{
sha256sum(filename)
}
\arguments{
\item{filename}{file to read}
}
\value{
sha256sum
}
\description{
reads a file and computes its sha256sum.
}
