% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{update metadata of a dataset}
\usage{
update_metadata(hash, meta)
}
\arguments{
\item{hash}{hash of the file that should be updated}

\item{meta}{metadata object, e.g. built with build_metadata, or from downloading existing}
}
\description{
update metadata of a dataset
}
\examples{
meta <- metadata(hash)
meta$share <- "public"
update_metadata(hash, meta)
}
