% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{validate_token}
\alias{validate_token}
\title{validates the token against datatomb
throws if token is invalid and sets user and admin fields in pkg.env}
\usage{
validate_token(token)
}
\arguments{
\item{token}{access token as obtained by, e.g., auth.spang-lab.de}
}
\value{
nothing
}
\description{
validates the token against datatomb
throws if token is invalid and sets user and admin fields in pkg.env
}
