% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{build_metadata}
\alias{build_metadata}
\title{internal function to check and complete metadata.}
\usage{
build_metadata(
  filename = NA,
  name = NA,
  tags = c(),
  projectname = NA,
  description = NA,
  parents = c(),
  data = list(),
  share = "internal"
)
}
\arguments{
\item{filename}{filename of the data file}

\item{...}{metadata fields}
}
\description{
internal function to check and complete metadata.
}
