% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{delete a data set remotely on datatomb}
\usage{
delete(hash, quiet = FALSE, force = FALSE)
}
\arguments{
\item{hash}{hash to delete}

\item{quiet}{if set to TRUE, don't ask before deletion}

\item{force}{if set to true, allow deletion of datasets that were not uploaded using diggeR}
}
\value{
nothing
}
\description{
delete a data set remotely on datatomb
}
