% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\title{upload a dataset to datatomb}
\usage{
upload(file, ...)
}
\arguments{
\item{file}{file to upload (on the file system)}

\item{...}{metadata as key-value pairs. see build_metadata}
}
\value{
hash of the uploaded file
}
\description{
upload a dataset to datatomb
}
\examples{
upload(file="datafile.h5", tags=c("testtag", "anothertesttag"), share="private")

upload(file="datafile.h5", description="a test file", projectname="myproject",
       tags=c("testtag", "anothertesttag"), share="private", data=list(additional="metadata"))

}
